Ref spec #389
